Move / rename of jsapi Picker utils (#1909)
