+++ packages/jsapi-components/src/spectrum/Picker/index.ts
@@ -1,2 +1,1 @@
 export * from './Picker';
-export * from './PickerUtils';
