+++ packages/jsapi-components/src/spectrum/utils/itemUtils.test.tsx
@@ -1,6 +1,6 @@
 import { type dh } from '@deephaven/jsapi-types';
 import { TestUtils } from '@deephaven/utils';
-import { getPickerKeyColumn, getPickerLabelColumn } from './PickerUtils';
+import { getItemKeyColumn, getItemLabelColumn } from './itemUtils';
 
 const { createMockProxy } = TestUtils;
 
@@ -22,27 +22,27 @@ beforeEach(() => {
   expect.hasAssertions();
 });
 
-describe('getPickerKeyColumn', () => {
+describe('getItemKeyColumn', () => {
   it.each([
     ['keyColumn', keyColumn],
     [undefined, columns[0]],
   ])(
     'should return the given key column or fallback to the first column',
     (keyColumnName, expectedColumn) => {
-      const actual = getPickerKeyColumn(table, keyColumnName);
+      const actual = getItemKeyColumn(table, keyColumnName);
       expect(actual).toBe(expectedColumn);
     }
   );
 });
 
-describe('getPickerLabelColumn', () => {
+describe('getItemLabelColumn', () => {
   it.each([
     ['labelColumn', labelColumn],
     [undefined, keyColumn],
   ])(
     'should return the given label column or fallback to the key column',
     (labelColumnName, expectedColumn) => {
-      const actual = getPickerLabelColumn(table, keyColumn, labelColumnName);
+      const actual = getItemLabelColumn(table, keyColumn, labelColumnName);
       expect(actual).toBe(expectedColumn);
     }
   );
